+++ docs/changelog.md
@@ -9,6 +9,17 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/firefox-android/blob/main/android-components/plugins/dependencies/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/firefox-android/blob/main/android-components/.config.yml)
 
+* **browser-storage-sync**:
+* **feature-awesomebar**
+* **feature-syncedtabs**
+  * 🆕 [bug #1800268](https://bugzilla.mozilla.org/show_bug.cgi?id=1800268) New autocompelte providers for bookmarks, local tabs or synced tabs that can be set for `ToolbarAutocompleteFeature`.
+
+* **feature-toolbar**
+  * ⚠️ **This is a breaking change**: `ToolbarAutocompleteFeature` has a new API for updating at any time `AutocompleteProvider` (add or remove any of them) individually or in bulk. This change allows supporting any instance and any number of autocomplete providers and optionally query for new autocomplete results when providers change. [bug #1800268](https://bugzilla.mozilla.org/show_bug.cgi?id=1800268)
+
+* **feature-awesomebar**
+  * 🆕 [bug #1800268](https://bugzilla.mozilla.org/show_bug.cgi?id=1800268) A new optional lambda `externalUrlFilter` has been added to the constructors of some AwesomeBar suggestions providers. This will allow for external url filtering of the provided suggestions.
+
 * **browser-state**, **feature-search**
   * Added a new parameter `isGeneral` to `SearchEngine` to specify whether or not the search engine is a general search engine (eg, provides broad search results). Search engines read from storage will now have this parameter set based on a list of general search engines. [bug #1804594](https://bugzilla.mozilla.org/show_bug.cgi?id=1804594)
   * Added Selector `BrowserState.findNormalOrPrivateTabByUrlIgnoringFragment` to match urls ignoring the fragment/anchor of the url, allowing `SelectorAddUseCase` to use this functionality.